app_server/pyproject.toml
[project]
name = "app-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "fastmcp>=0.1.0",
    "httpx>=0.27.0",
    "jinja2>=3.1.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.1",
    "uvicorn[standard]>=0.29.0",
]
